fix(desktopui): handle commit deleted event
previously, commit created and deleted were handled by the same event
as the handler just updates the branches.

however, we now show a notification with info on the event.
branch name is only returned on commit created, so a diff handler is
now needed for commit deleted. this is the fix!

tested in both Rhino & Revit

closes #371