fix(desktopui): don't reload accts on toggle
in stream create dialog:
refreshing the accounts when toggling the accounts dropdown was
deselecting the selected acct. resetting it will always refresh the ui
which is an unpleasant experience, so i've opted to just take it out.
we have a refresh button and in theory the accts should reload anyway
if you close and reopen the dialog.

closes #474

also no acct issue here shouldn't crash anymore thanks to #480