fix(revit): override units for internal duct dimensions
in imperial, these are in inches while the model units are in feet.
this meant that eg 12 was set as the diameter for a duct which would be
received as 12 feet instead of 12 inches.
as far as i can tell, this only affects mechanical. however keep an eye
out for other possible scenarios like this where we are setting the
value of a parameter to the actual speckle object