fix: Correctly process aggregate models that promote to the root node.
When aggregating models, Navisworks aggregates some formats to be at the root node and in others, the root node is void and the first child is the quasi root node. This adds a pseudoId just for root.

TODO: detect both cases and only use this root node pseudoId when needed.