Release 2.12 additions for Navisworks Alpha (#2095)
* Expanding the Conversion State possibilities to Enum from bool

In theory, further states could be added and useful

* alpha3 version flag

* fix: Kit loading logic and instantiation

* fix: smarter conversion state tracking

* fix: Model Orientation XY problem with non-architectural models

ElevationMode implementation checks for the current UP vector for the document and overrides YZ axis flipping on the basis that Speckle assumes XY ground plane with Z elevation.

* fix: Correctly process aggregate models that promote to the root node.

When aggregating models, Navisworks aggregates some formats to be at the root node and in others, the root node is void and the first child is the quasi root node. This adds a pseudoId just for root.

TODO: detect both cases and only use this root node pseudoId when needed.

* fix: tweak to the YZ issue fix #2015

TODO: detect for the handedness of the model to correctly apply the negative Z fix only when necessary.

* fix: selection logic improvements

distinction between selected object count, objects to convert count, and converted counts

* fix: further code legibility improvements

* Prevent potentially empty displayValue from being sent

Ideally this should mean that a later generic check would be that any leaf node that doesn't have a display value should also not be sent

* DUI Changes for Advanced Settings

* Implemented Send Settings for Origin Transformation

Closes #2085

* Addition of units for the Coordinate Transform.

Added a Unit dictionary I don't know if I'll use yet.

* Modified the TreeViewFilter to accept SelectionMode Binding

Defaults to Multiple + Toggle Select

* Populate a treeview filter with Saved Viewpoints

TODO: work out an id scheme for SavedViews all GUIDs are zeroed

* Disabled addition of bbox. Fixes: #2050

Commented out. Method intact for a rainy day

* Null checking on DocumentDatabase - Fixes #2091

* chore (Navisworks): Adds file links to Core .editorConfigs

Adopts the Speckle Core coding standards moving forward

* Expands the `GetPseudoId` helper to convert more input types

This primarily adds support for the `IndexPath` property of `SavedViewpoints` and `SelectionSets`. The rest is other possible entry types for conversion

* feature(Navisworks): Added Lines & Polylines as supported geometries

TODO: Polylines are a collection of Lines, Polylines are not tested or set as closed.
TODO: Splines and Circles not yet supported.

* fix: Sending no valid objects now safely cancels and returns control

* SavedView implementation scaffold

TODO: SavedView as element source

* Implemented Commit Views and Views as Object Source

TODO: Canonical view type views are functional, Some residual view orientation quirks with wide FOVs

* chore: remove unused dependencies

---------

Co-authored-by: Matteo Cominetti <matteo@speckle.systems>