Add "Retry Send" button to Navisworks ribbon
This commit adds a new button to the Navisworks ribbon that allows users to retry sending the last complete conversion to Speckle. The new button is added in Ribbon.xaml and Ribbon.xaml.cs, and its functionality is implemented in RibbonHandler.cs. The code also includes some minor changes such as renaming a command shortcut and adding comments for clarity.