feat(Navisworks): Conversion Cache (#2682)
* Add "Retry Send" button to Navisworks ribbon

This commit adds a new button to the Navisworks ribbon that allows users to retry sending the last complete conversion to Speckle. The new button is added in Ribbon.xaml and Ribbon.xaml.cs, and its functionality is implemented in RibbonHandler.cs. The code also includes some minor changes such as renaming a command shortcut and adding comments for clarity.

* Add caching for last conversion and commit

This commit adds caching for the last conversion and commit objects to improve performance when retrying a failed send operation. The cached objects are used in the new `RetryLastConversionSend` method, which allows users to retry sending data without having to re-convert everything from scratch.

* Refactor: Clear cached data after successfully creating a commit

This change clears the cachedCommit and cachedConversion variables after creating a commit in ConnectorNavisworksBindings.Send.cs. This ensures that the latest data is always used and avoids any potential issues with stale data.

* Implement RetryLastConversionSend()

- Added new using statements for Models and Utilities.
- Added a new method RetryLastConversionSend() to ConnectorBindingsNavisworks class.
- Modified SendConvertedObjectsToSpeckle() and CreateCommit() methods in ConnectorBindingsNavisworks class to use the cached conversion objects.
- Added a new field _cachedState to store the state of the stream before sending data to Speckle.
- Modified ExecuteCommand() method in RibbonHandler class to call RetryLastConversionSend() method from Bindings object instead of directly calling it.
- Created a new field Bindings in SpeckleNavisworksCommandPlugin class.

* Adds cache cancellation for Document Change and new Sends

- Remove unused variables
- Clear cached objects when necessary

* Add isRetrying flag to ConnectorBindingsNavisworks and handle it in ConnectorBindingsNavisworks.Send.cs

This commit adds a new boolean flag called isRetrying to the ConnectorBindingsNavisworks class. It also handles this flag in the Send method of the same class, where it throws an exception if isRetrying is false and debug mode is enabled.

* Add retry button to Ribbon and Resources

This commit adds a retry button to the Ribbon in ConnectorNavisworks. It also includes two new resources, "retry16.ico" and "retry32.ico", which are used as icons for the new button. Additionally, some code was commented out that threw an exception when not in debug mode.

* add comments

* Improve caching and add cache persistence

This commit refactors the ConnectorNavisworksBindings.cs file to improve caching by using a list of Base objects for cached conversions. It also adds a new Ribbon button toggle that allows users to turn on or off cache persistence, which determines whether converted objects are held in memory after sending them to the Stream and Branch specified by the user. The changes include: 
modifying CachedConversion state check, 
add CachedConvertedElements container object
add TurnPersistCacheOn, and TurnPersistCacheOff Commands.

* Refactor caching and retrying logic in ConnectorNavisworksBindings

This commit refactors the caching and retrying logic in ConnectorNavisworksBindings. It removes the _isRetrying flag and replaces it with comments to disable code blocks that are no longer needed. Additionally, it fixes a bug where CachedConvertedElements were not being assigned properly when retrieving from the cache. Finally, it updates an ExtendedToolTip string to fix a typo.