use internal property names
This commit refactors the code to use internal property names instead of display names if so set in the converter settings. It also adds a new setting for using internal property names and updates the `ProcessPropertyCategory` method to handle this change. Finally, it updates the `InitializeManagerOptionsForSend` method to seed the `SettingsManager` with options for conversion.