better handle autosave settings during Send operation
This commit refactors the Autosave.cs file to handle autosave settings during a Send operation. The code now correctly checks if autosave was turned on or off at the time of the send operation and sets it accordingly. If it was turned off, it will remain off until manually turned back on. If it was already on, it will be turned back on after the send operation is complete.