feat(Navisworks): Options to include Internal Properties and Use Internal Property Names (#2697)
* Add NavisworksOptionsManager class with Autosave set/get

This commit adds a new Autosave.cs file to the ConnectorNavisworks project. The file contains a NavisworksSettingsManager class that handles updating and restoring the auto-save setting in Navisworks.

It is modelled as a Partial class as a stub for potentially handling other Host Application Global Options interop.

* Adds Settings Options allowing to always show regardless of internal settings

* Refactor ConnectorNavisworksBindings.cs and ConnectorNavisworksBindings.Send.cs
- Add NavisworksOptionsManager to handle settings
- Remove unused variables and methods
- Add new settings for commit properties and internal properties

* Add calls to Property names setting

it resets the options manager after completion or cancellation of the process.

* use internal property names

This commit refactors the code to use internal property names instead of display names if so set in the converter settings. It also adds a new setting for using internal property names and updates the `ProcessPropertyCategory` method to handle this change. Finally, it updates the `InitializeManagerOptionsForSend` method to seed the `SettingsManager` with options for conversion.

* Refactor element name handling in ConverterNavisworks.ToSpeckle

This commit refactors the way element names are handled in the ConverterNavisworks.ToSpeckle class. Specifically, it changes how resolvedName is used and adds a ToString() method to an object.

* element name resolution

This commit changes the way element names are resolved in the ToSpeckle method of ConverterNavisworks. Specifically, it updates how parent elements are found and how the "name" property is set for the @base object.

* better handle autosave settings during Send operation

This commit refactors the Autosave.cs file to handle autosave settings during a Send operation. The code now correctly checks if autosave was turned on or off at the time of the send operation and sets it accordingly. If it was turned off, it will remain off until manually turned back on. If it was already on, it will be turned back on after the send operation is complete.