fix(Navisworks): Handle Incomplete Viewpoint Definitions (#2706)
* Refactor camera variables and add error handling for null focal distance

The code changes refactor the camera variables to be more descriptive and adds error handling for when a selected view's viewpoint has no focal distance set. The commit message follows an informal tone, lists significant changes, and is concise.

* Extend Invocation of SavedViewpoint to Current

The SavedViewpoint is retrieved and then assigned to current on the same thread.

* Refactor error handling for Navisworks converter

Improved error handling in the Navisworks converter by adding support for a new exception type. Updated logging message to reflect that there are now two possible exceptions that can be thrown when a selected view's viewpoint has no focal distance set.