feat(Navisworks): Option to coalesce all data from First Object to the Geometry Node (#2724)
* Add property stack to geometry node

The `AddPropertyStackToGeometryNode` method is introduced, which takes in the dictionary of converted elements, a model item, and a base node. It retrieves the FirstObjectAncestor and trims the ancestors accordingly. Then, it constructs the property stack by selecting relevant properties from each ancestor and grouping them by category. Finally, it assigns the property stack to the base node.

* Removes the unnecessary assignment of a default value to the zoomValue variable.

* Refactor BuildNestedObjectHierarchy method to accept a new dictionary structure and add logic for handling GeometryNode types.

- BuildNestedObjectHierarchy receives converted objects rather than a remapped dictionary.
- Added logic to convert the new dictionary structure into the previously passed in convertedDictionary format.
- Implemented logic to handle GeometryNode types by adding properties to the FirstObject classification.

* Change ConversionState enum access modifier to public

- Change the access modifier of the ConversionState enum from internal to public in the Constants.cs file.

* Refactor code for building a family tree in ConnectorNavisworksBindings.Send.cs

- Simplify code by directly passing the converted dictionary to BuildNestedObjectHierarchy method
- Update commitObject.elements assignment accordingly

* Map propertyStack back to properties

* include state parameter in BuildNestedObjectHierarchy method

The commit updates the code in ConnectorNavisworksBindings.Send.cs file. The change includes adding a state parameter to the BuildNestedObjectHierarchy method call

* Add coalesce data functionality to SelectionHandler

This commit adds the ability to coalesce data from the first object ancestor in the SelectionHandler class. If the "coalesce-data" checkbox is checked and the full tree is not selected, relevant parents are added to ensure that data is coalesced correctly.

* Add support for coalescing data in the BuildNestedObjectHierarchy method

This commit adds support for coalescing data in the `BuildNestedObjectHierarchy` method. It introduces a new parameter `streamState` and checks if the "coalesce-data" setting is enabled before adding properties to GeometryNodes.

* feat: Add new checkbox setting for coalescing data

This commit adds a new checkbox setting called "Coalesce Data from First Object to Geometry". When enabled, all properties from the Geometry up the tree to the next First Object will be coalesced into the Geometry Node. This feature can be useful for downstream data analysis.

* comment out the assignment of `propertyStack` to `baseNode`