De-coupled bridge and bindings
not entirely happy with it but the binding and bridge are not circular dependencies, and they can now be bound together. This has the neat effect of resolving everything on the right context and it loads up.

The bindings do appear to be broken, so I'm going to need to dig into that.