Initial state for DUI3/Alpha development kick-off 🚀 (#3228)
* feat(dui3): Adds DUI3 projects and current implementation

This is coming from dui3000 branch!!! It was squashed and selectively comited here.

* chore(core): Flagged some classes as ComVisible

* chore(acad/rhino/revit): Temporary modification of converters to work with DUI3

This should be revised/reverted before finalising DUI3

* Fix(Styling): fix remaining styling issues after squash

* Remove unused using statements

* Chore(Dui3): Add slnf for DUI3 projects (#3088)

Add solution filter for DUI3 projects

* Align with core changes

* Feat (DUI3): ArcGIS POC (#3156)

* Bump System.Threading.Tasks.Dataflow to version 6

- It was causing assembly load error on ArcGIS

* Init ArcGIS project template for DUI3

* Fix warnings as much as possible

* Fix solution filter merge conflicts

* Finding DI Shape (Broken commit)

This commit is broken but finding the shape is a bit painful and slow, don't want to lose progress

* Partial Assembly Setup & DI

as per above

* DUI3 DI solution filter

some more assembly/solution org

* Standing up the plugin

Plugin is loading with Autofac dependency. Carving in other dependencies ahs begun...

* Fix (DUI3): CNX-8930 Add UseWPF prop to AutoCAD projects (#3178)

Add UseWPF prop to AutoCAD projects

* Panel Appearing

Have the CefSharp panel appearing

* feat(dui3): DUI3 project copy paste merge

* feat(dui3): DUI3 Connectors project copy paste merge

* feat(dui3): highlighting an empty model sender card throws an error

* feat(dui3): quite a few fixes, do sync with with dui3 web

* Binding work in Progress

DI of bindings and dependencies

* Bindings injected

Bindings are now injected but need to unpick the relationship between the bridge and the binding (given they know about each other)

* feat(dui3): acad23 send

* feat(dui3): acad23 receive + layers

* feat(dui3): acad receive, highlight receiver cards & a few more tweaks

* feat(dui3): fixes a bug in the selection bindings; fixes a bg in the get objects from document function that used a risky method (replaced with a more gentle one)

* feat(dui3): revert unintentional proj change

* DI Bindings

With DI in the mix, the bindings are creating associations with Revit from within different contexts and this causing some explosions.

Ideally the eventing methods should bind loosely and be decoupled from the implementation. Gotta rethink a bit

* De-coupled bridge and bindings

not entirely happy with it but the binding and bridge are not circular dependencies, and they can now be bound together. This has the neat effect of resolving everything on the right context and it loads up.

The bindings do appear to be broken, so I'm going to need to dig into that.

* Bindings Injected

Finally these are being injected neatly, so most statics and discovery made that bit easier. The bindings also appear to be working, calling into the C# code from Revit on selection change.
Next up: get the calling into each binding, get some account context in there, start in on the kits or lack thereof :D

* Binding NOT working

Seems on the browser side it's still broken from a binding POV

* Binding now really DOES work

Binding had the appearance of working, but I failed in my refactor to understand that some of the changes - Method -> Property change of return type, etc... were breaking the symmetry required on the JS side as well and things stop working. Now working pretty fine.

Next steps: bind some unnecessary changes and re-instate DI for Panel. Check this works and the look at Send operation and Kits

* Re-instated the DI for CefSharpPanel

The re-instated DI still works, now the contract is being honoured

* Updated Id

Id -> ModelCardId

* DUI3/Dev -> DI DUI3

Merged main DUI3 branch

* Remove cefsharp 121.3.7 version existence on project

* TypeLoadException was blowing up the deserializing of the SenderModelCard

* Interim fix of DiscriminatedObjectConverter

* Statics clean-up

Removed the obvious statics that remain in the DUI area and converted some to consts

* Binding re-instate

Re-instate the binding for document changed

* Library Additions

Added Objects, Converters related libs and moved over 'enough' to add Floor.cs

* Minor changes...

to namespaces

* DI Converter Shape

Have added some nascent interfaces and a class for discovering the converters and wiring this into the Send binding

* Disposing & Factories

Organised some factories and dispose pattern ready for send operation

* Logging woes + DI Problems with lifetimes

The microsoft logging extensions did not want to load here, but have nailed the versions. Also had issues with DI of factory but resolved

* Type discovery

Discovering types dynamically

* Tweaks

Tweaks

---------

Co-authored-by: oguzhankoral <oguzhankoral@gmail.com>
Co-authored-by: Oğuzhan Koral <45078678+oguzhankoral@users.noreply.github.com>
Co-authored-by: Jedd Morgan <45512892+JR-Morgan@users.noreply.github.com>
Co-authored-by: Ian Hawley <ian@speckle.systems>
Co-authored-by: Dimitrie Stefanescu <didimitrie@gmail.com>